pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ops-pilot"
version = "0.1.0"
description = "Agentic CI/CD incident responder — monitors pipelines, triages failures, generates fixes, notifies teams"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "anthropic>=0.25.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "httpx>=0.27.0",
]

[project.scripts]
ops-pilot = "agents.monitor_agent:main"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=agents --cov=shared --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
omit = ["tests/*", "demo/*"]

[tool.hatch.build.targets.wheel]
packages = ["agents", "shared"]
